requirements.txt
discord.py==2.4.0
requests==2.32.3
apscheduler==3.11.0
tzlocal==5.3.1
matplotlib==3.10.8
psutil==7.2.2
docker==7.1.0